Extract delegated spend parsing
