+++ crates/chia-sdk-driver/src/clear_signing.rs
@@ -1,17 +1,19 @@
+mod delegated_spend;
 mod facts;
 mod inner_spend;
 mod linked_spend;
 mod requested_payments;
 mod signing;
-mod vault;
 mod vault_message;
+mod vault_spend;
 mod vault_transaction;
 
+pub use delegated_spend::*;
 pub use facts::*;
 pub use inner_spend::*;
 pub use linked_spend::*;
 pub use requested_payments::*;
 pub use signing::*;
-pub use vault::*;
 pub use vault_message::*;
+pub use vault_spend::*;
 pub use vault_transaction::*;
