+++ crates/chia-sdk-driver/src/clear_signing/vault_spend.rs
